Added the synchonization of CRUDE and load/save
Added Feature in this patch:
1.synched properly of load and save(didn't scrutinize so there might be bugs, check later)
2. The cats are bought and placed with first in first add method

Here are the list on what-to-do now:

Additional Features that is needed:
1. Add customization of the frame background like changing the house theme.
2. Add movable objects that can be bought(or Cats if possible).
3. DELETE FEATURE(IMPOSSIBLE)

Committed on 8:47 of 16th of December 2023.

PS: The CRUDE is halfway done! idk if i want to put delete feature because it is so complicated now. This patch is buggy but i need to rest for now.Also, there are a lot of optimization issues and redundancy but if i remove them something might not work.

-SerenicuS