Optimization and Debug Statements added and currently being worked on, to make sure that adding new features will not be a hassle.
THIS PATCH IS THE MOST BUGGY PATCH EVER, PLEASE FIX THIS ASAP.

TODO:  ADD EVERY FUNCTIONS WITH PRINT STATEMENTS TO DEBUG AND EASILY ADD MORE FEATURES. CHECK THE CODE TO SEE WHERE IS THE LAST FUNCTION THAT HAS REVISED PRINT STATEMENTS/DEBUGS STATEMENTS

FIXME:
1. A code here that is still not found is adding a phantom value that is being added in arrangementOfCats for some reason and it makes the add of new cats weird because the 2nd catimage will go to the 3rd catimage's location.

2. After buying a cat and bought another one, the CatInfo label showing the cat's image before being bought is not the cat's own image and it is the previous cat that is bought. For some reason if you close that frame and click it again it will show the correct label, what annoying bug.

Committed on 10:52pm of 17th of December 2023.

PS: Man, this is one of the most buggy update ever. I need to fix this asap so i can start creating the soft copy for this.

-SerenicuS