Some Bugs are fixed.
This is a much better update, fixed the synchronization issues but has phantom bug that cannot be found for some reason when I exited the program, leading to load and save feature not in sync

TODO: If possible, try to fix the exit load and save bug

FIXME more details:
When I dipose the frame, the load and save progress and its contents are synchronized. However, when I try to exit the frame, it will lead to some phantom bugs that I cannot really fix even though my program has literally Debug Statements all over it. If i can't fix it right now, i might as well not continue fixing it.

Committed on 12:34pm of 18th of December 2023.

PS: One of the hardest bugs to ever fix and find.

-SerenicuS