This Program is now clean,
This is the pre-final update, all is good and much of it are now synchronized. Although I still didn't scrutinize it to the point that I am going to hate it. Still it looks clean now.

Suggestions:
1. Add a DELETE BUTTON.

Committed on 2:01pm of 18th of December 2023.

PS: Fixed the Load and Save Bug after exiting the program. Whether I or not add a delete button will depends on the time remaining.

-SerenicuS